+++ packages/firebase_auth/firebase_auth/example/test_driver/instance_e2e.dart
@@ -178,8 +178,6 @@ void runInstanceTests() {
         await FirebaseAuth.instance.currentUser!
             .updateDisplayName('updatedName');
 
-        await FirebaseAuth.instance.currentUser!.reload();
-
         expect(
           FirebaseAuth.instance.currentUser!.displayName,
           equals('updatedName'),