+++ tests/test_driver/cloud_functions/cloud_functions_e2e.dart
@@ -3,12 +3,11 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:cloud_functions/cloud_functions.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:drive/drive.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/foundation.dart';
 
 import '../firebase_default_options.dart';
-
 import 'sample_data.dart' as data;
 
 String kTestFunctionDefaultRegion = 'testFunctionDefaultRegion';