+++ packages/cloud_firestore_odm/cloud_firestore_odm_generator/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0-dev.29
+
+ - **FIX**: bump minimum analyzer version ([#9493](https://github.com/firebase/flutterfire/issues/9493)). ([5137a646](https://github.com/firebase/flutterfire/commit/5137a6469fb57fb003757459222cb6c4e39fb0f8))
+ - **FEAT**: Add support using Freezed classes as collection models ([#9483](https://github.com/firebase/flutterfire/issues/9483)). ([ce238f71](https://github.com/firebase/flutterfire/commit/ce238f713b250f523890b9e7e42d395f433ed80f))
+
 ## 1.0.0-dev.28
 
  - Update a dependency to the latest release.