+++ packages/cloud_functions/cloud_functions_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.3.5
+
+ - Update a dependency to the latest release.
+
 ## 4.3.4
 
  - Update a dependency to the latest release.